+++ src/main/java/me/comu/exeter/commands/image/CompressImageCommand.java
@@ -13,57 +13,91 @@ import java.io.File;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.TimeUnit;
 
 public class CompressImageCommand implements ICommand {
 
 
     @Override
     public void handle(List<String> args, GuildMessageReceivedEvent event) {
+        if (Wrapper.beingProcessed) {
+            event.getChannel().sendMessage("An image is already being processed, please wait.").queue();
+            return;
+        }
         if (event.getMessage().getAttachments().isEmpty()) {
             if (args.isEmpty()) {
                 event.getChannel().sendMessage("Please insert an image link to manipulate").queue();
                 return;
             }
+            Wrapper.beingProcessed = true;
             event.getChannel().sendMessage("`Processing Image...`").queue(message -> {
                 try {
                     int random = new Random().nextInt(1000);
                     int newRandom = new Random().nextInt(1000);
                     Wrapper.saveImage(args.get(0), "cache", "image" + random);
                     File file = new File("cache/image" + random + ".png");
-                    Image img;
-                    BufferedImage tempPNG;
-                    File newFilePNG;
-                    img = ImageIO.read(file);
-                    double aspectRatio = (double) img.getWidth(null)/(double) img.getHeight(null);
-                    tempPNG = resizeImage(img, 1080, (int) (3000/aspectRatio));
-                    newFilePNG = new File("cache/image" + newRandom + ".png");
-                    ImageIO.write(tempPNG, "png", newFilePNG);
-                    message.delete().queue();
-                    event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                    BufferedImage image = ImageIO.read(file);
+                    CompletableFuture.supplyAsync(() -> image)
+                            .thenApply(img -> resizeImage(img, 1080, 3000))
+                            .completeOnTimeout(null, 10, TimeUnit.SECONDS)
+                            .thenAccept(processedImage -> {
+                                if (processedImage == null) {
+                                    message.editMessage("Processing thread timed out.").queue();
+                                    Config.clearCacheDirectory();
+                                    Wrapper.beingProcessed = false;
+                                } else {
+                                    try {
+                                        File newFilePNG = new File("cache/image" + newRandom + ".png");
+                                        ImageIO.write(processedImage, "png", newFilePNG);
+                                        message.delete().queue();
+                                        event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                                        Wrapper.beingProcessed = false;
+                                    } catch (Exception ignored) {
+                                        message.editMessage("Something went wrong with processing the image").queue();
+                                        Wrapper.beingProcessed = false;
+                                    }
+                                }
+                            });
                 } catch (Exception ex) {
                     message.editMessage("Something went wrong with processing the image").queue();
+                    Wrapper.beingProcessed = false;
                 }
 
             });
         } else {
+            Wrapper.beingProcessed = true;
             event.getChannel().sendMessage("`Processing Image...`").queue(message -> {
                 try {
                     int random = new Random().nextInt(1000);
                     int newRandom = new Random().nextInt(1000);
                     Wrapper.saveImage(event.getMessage().getAttachments().get(0).getUrl(), "cache", "image" + random);
                     File file = new File("cache/image" + random + ".png");
-                    Image img;
-                    BufferedImage tempPNG;
-                    File newFilePNG;
-                    img = ImageIO.read(file);
-                    double aspectRatio = (double) img.getWidth(null)/(double) img.getHeight(null);
-                    tempPNG = resizeImage(img, 1080, (int) (3000/aspectRatio));
-                    newFilePNG = new File("cache/image" + newRandom + ".png");
-                    ImageIO.write(tempPNG, "png", newFilePNG);
-                    message.delete().queue();
-                    event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                    BufferedImage image = ImageIO.read(file);
+                    CompletableFuture.supplyAsync(() -> image)
+                            .thenApply(img -> resizeImage(img, 1080, 3000))
+                            .completeOnTimeout(null, 10, TimeUnit.SECONDS)
+                            .thenAccept(processedImage -> {
+                                if (processedImage == null) {
+                                    message.editMessage("Processing thread timed out.").queue();
+                                    Config.clearCacheDirectory();
+                                    Wrapper.beingProcessed = false;
+                                } else {
+                                    try {
+                                        File newFilePNG = new File("cache/image" + newRandom + ".png");
+                                        ImageIO.write(processedImage, "png", newFilePNG);
+                                        message.delete().queue();
+                                        event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                                        Wrapper.beingProcessed = false;
+                                    } catch (Exception ignored) {
+                                        message.editMessage("Something went wrong with processing the image").queue();
+                                        Wrapper.beingProcessed = false;
+                                    }
+                                }
+                            });
                 } catch (Exception ex) {
                     message.editMessage("Something went wrong with processing the image").queue();
+                    Wrapper.beingProcessed = false;
                 }
             });
         }
@@ -94,7 +128,7 @@ public class CompressImageCommand implements ICommand {
 
     @Override
     public String[] getAlias() {
-        return new String[]{"compressimage", "compressimg"};
+        return new String[]{"compressimage", "compressimg","compression"};
     }
 
     @Override
