+++ src/main/java/me/comu/exeter/commands/image/DistortImageCommand.java
@@ -13,17 +13,24 @@ import java.io.File;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.TimeUnit;
 
 public class DistortImageCommand implements ICommand {
 
 
     @Override
     public void handle(List<String> args, GuildMessageReceivedEvent event) {
+        if (Wrapper.beingProcessed) {
+            event.getChannel().sendMessage("An image is already being processed, please wait.").queue();
+            return;
+        }
         if (event.getMessage().getAttachments().isEmpty()) {
             if (args.isEmpty()) {
                 event.getChannel().sendMessage("Please insert an image link to manipulate").queue();
                 return;
             }
+            Wrapper.beingProcessed = true;
             event.getChannel().sendMessage("`Processing Image...`").queue(message -> {
                 try {
                     int random = new Random().nextInt(1000);
@@ -31,32 +38,36 @@ public class DistortImageCommand implements ICommand {
                     Wrapper.saveImage(args.get(0), "cache", "image" + random);
                     File file = new File("cache/image" + random + ".png");
                     BufferedImage image = ImageIO.read(file);
-                    for (int i = 0; i < image.getWidth(); i++) {
-                        for (int j = 0; j < image.getHeight(); j++) {
-                            if (i <= image.getWidth() && j <= image.getHeight()) {
-                                Color c = new Color(image.getRGB(i, j));
-                                int r = c.getRed();
-                                int g = c.getGreen();
-                                int b = c.getBlue();
-                                int a = c.getAlpha();
-                                Color rgb = new Color(r, g, b, a);
-                                if (Math.random() > 0.6)
-                                    image.setRGB(j, i, rgb.getRGB());
-                                else
-                                    image.setRGB(i, j, rgb.getRGB());
-                            }
-                        }
-                    }
-                    File newFilePNG = new File("cache/image" + newRandom + ".png");
-                    ImageIO.write(image, "png", newFilePNG);
-                    message.delete().queue();
-                    event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                    CompletableFuture.supplyAsync(() -> image)
+                            .thenApply(this::distortImg)
+                            .completeOnTimeout(null, 10, TimeUnit.SECONDS)
+                            .thenAccept(processedImage -> {
+                                if (processedImage == null) {
+                                    message.editMessage("Processing thread timed out.").queue();
+                                    Config.clearCacheDirectory();
+                                    Wrapper.beingProcessed = false;
+                                } else {
+                                    try {
+                                        File newFilePNG = new File("cache/image" + newRandom + ".png");
+                                        ImageIO.write(processedImage, "png", newFilePNG);
+                                        message.delete().queue();
+                                        event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                                        Wrapper.beingProcessed = false;
+                                    } catch (Exception ignored) {
+                                        message.editMessage("Something went wrong with processing the image").queue();
+                                        Wrapper.beingProcessed = false;
+                                    }
+                                }
+                            });
+
                 } catch (Exception ignored) {
                     message.editMessage("Something went wrong with processing the image").queue();
+                    Wrapper.beingProcessed = false;
                 }
 
             });
         } else {
+            Wrapper.beingProcessed = true;
             event.getChannel().sendMessage("`Processing Image...`").queue(message -> {
                 try {
                     int random = new Random().nextInt(1000);
@@ -64,31 +75,55 @@ public class DistortImageCommand implements ICommand {
                     Wrapper.saveImage(args.get(0), "cache", "image" + random);
                     File file = new File("cache/image" + random + ".png");
                     BufferedImage image = ImageIO.read(file);
-                    for (int i = 0; i < image.getWidth(); i++) {
-                        for (int j = 0; j < image.getHeight(); j++) {
-                            Color c = new Color(image.getRGB(i, j));
-                            int r = c.getRed();
-                            int g = c.getGreen();
-                            int b = c.getBlue();
-                            int a = c.getAlpha();
-                            Color rgb = new Color(r, g, b, a);
-                            if (Math.random() > 0.6)
-                                image.setRGB(j, i, rgb.getRGB());
-                            else
-                                image.setRGB(i, j, rgb.getRGB());
-                        }
-                    }
-                    File newFilePNG = new File("cache/image" + newRandom + ".png");
-                    ImageIO.write(image, "png", newFilePNG);
-                    message.delete().queue();
-                    event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                    CompletableFuture.supplyAsync(() -> image)
+                            .thenApply(this::distortImg)
+                            .completeOnTimeout(null, 10, TimeUnit.SECONDS)
+                            .thenAccept(processedImage -> {
+                                if (processedImage == null) {
+                                    message.editMessage("Processing thread timed out.").queue();
+                                    Config.clearCacheDirectory();
+                                    Wrapper.beingProcessed = false;
+                                } else {
+                                    try {
+                                        File newFilePNG = new File("cache/image" + newRandom + ".png");
+                                        ImageIO.write(processedImage, "png", newFilePNG);
+                                        message.delete().queue();
+                                        event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                                        Wrapper.beingProcessed = false;
+                                    } catch (Exception ignored) {
+                                        message.editMessage("Something went wrong with processing the image").queue();
+                                        Wrapper.beingProcessed = false;
+                                    }
+                                }
+                            });
                 } catch (Exception ex) {
                     message.editMessage("Something went wrong with processing the image").queue();
+                    Wrapper.beingProcessed = false;
                 }
 
             });
         }
         Config.clearCacheDirectory();
+    }
+
+    private BufferedImage distortImg(BufferedImage image) {
+        for (int i = 0; i < image.getWidth(); i++) {
+            for (int j = 0; j < image.getHeight(); j++) {
+                if (i <= image.getWidth() && j <= image.getHeight()) {
+                    Color c = new Color(image.getRGB(i, j));
+                    int r = c.getRed();
+                    int g = c.getGreen();
+                    int b = c.getBlue();
+                    int a = c.getAlpha();
+                    Color rgb = new Color(r, g, b, a);
+                    if (Math.random() > 0.6)
+                        image.setRGB(j, i, rgb.getRGB());
+                    else
+                        image.setRGB(i, j, rgb.getRGB());
+                }
+            }
+        }
+        return image;
     }
