+++ src/main/java/me/comu/exeter/commands/image/GrayScaleImageCommand.java
@@ -12,17 +12,24 @@ import java.io.File;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.TimeUnit;
 
 public class GrayScaleImageCommand implements ICommand {
 
 
     @Override
     public void handle(List<String> args, GuildMessageReceivedEvent event) {
+        if (Wrapper.beingProcessed) {
+            event.getChannel().sendMessage("An image is already being processed, please wait.").queue();
+            return;
+        }
         if (event.getMessage().getAttachments().isEmpty()) {
             if (args.isEmpty()) {
                 event.getChannel().sendMessage("Please insert an image link to manipulate").queue();
                 return;
             }
+            Wrapper.beingProcessed = true;
             event.getChannel().sendMessage("`Processing Image...`").queue(message -> {
                 try {
                     int random = new Random().nextInt(1000);
@@ -30,30 +37,35 @@ public class GrayScaleImageCommand implements ICommand {
                     Wrapper.saveImage(args.get(0), "cache", "image" + random);
                     File file = new File("cache/image" + random + ".png");
                     BufferedImage image = ImageIO.read(file);
-                    int width = image.getWidth();
-                    int height = image.getHeight();
-                    for (int y = 0; y < height; y++) {
-                        for (int x = 0; x < width; x++) {
-                            int p = image.getRGB(x, y);
-                            int a = (p >> 24) & 0xff;
-                            int r = (p >> 16) & 0xff;
-                            int g = (p >> 8) & 0xff;
-                            int b = p & 0xff;
-                            int avg = (r + g + b) / 3;
-                            p = (a << 24) | (avg << 16) | (avg << 8) | avg;
-                            image.setRGB(x, y, p);
-                        }
-                    }
-                    File newFilePNG = new File("cache/image" + newRandom + ".png");
-                    ImageIO.write(image, "png", newFilePNG);
-                    message.delete().queue();
-                    event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                    CompletableFuture.supplyAsync(() -> image)
+                            .thenApply(this::grayScaleImg)
+                            .completeOnTimeout(null, 10, TimeUnit.SECONDS)
+                            .thenAccept(processedImage -> {
+                                if (processedImage == null) {
+                                    message.editMessage("Processing thread timed out.").queue();
+                                    Config.clearCacheDirectory();
+                                    Wrapper.beingProcessed = false;
+                                } else {
+                                    try {
+                                        File newFilePNG = new File("cache/image" + newRandom + ".png");
+                                        ImageIO.write(processedImage, "png", newFilePNG);
+                                        message.delete().queue();
+                                        event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                                        Wrapper.beingProcessed = false;
+                                    } catch (Exception ignored) {
+                                        message.editMessage("Something went wrong with processing the image").queue();
+                                        Wrapper.beingProcessed = false;
+                                    }
+                                }
+                            });
                 } catch (Exception ignored) {
                     message.editMessage("Something went wrong with processing the image").queue();
+                    Wrapper.beingProcessed = false;
                 }
 
             });
         } else {
+            Wrapper.beingProcessed = true;
             event.getChannel().sendMessage("`Processing Image...`").queue(message -> {
                 try {
                     int random = new Random().nextInt(1000);
@@ -61,33 +73,55 @@ public class GrayScaleImageCommand implements ICommand {
                     Wrapper.saveImage(event.getMessage().getAttachments().get(0).getUrl(), "cache", "image" + random);
                     File file = new File("cache/image" + random + ".png");
                     BufferedImage image = ImageIO.read(file);
-                    int width = image.getWidth();
-                    int height = image.getHeight();
-                    for (int y = 0; y < height; y++) {
-                        for (int x = 0; x < width; x++) {
-                            int p = image.getRGB(x, y);
-                            int a = (p >> 24) & 0xff;
-                            int r = (p >> 16) & 0xff;
-                            int g = (p >> 8) & 0xff;
-                            int b = p & 0xff;
-                            int avg = (r + g + b) / 3;
-                            p = (a << 24) | (avg << 16) | (avg << 8) | avg;
-                            image.setRGB(x, y, p);
-                        }
-                    }
-                    File newFilePNG = new File("cache/image" + newRandom + ".png");
-                    ImageIO.write(image, "png", newFilePNG);
-                    message.delete().queue();
-                    event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                    CompletableFuture.supplyAsync(() -> image)
+                            .thenApply(this::grayScaleImg)
+                            .completeOnTimeout(null, 10, TimeUnit.SECONDS)
+                            .thenAccept(processedImage -> {
+                                if (processedImage == null) {
+                                    message.editMessage("Processing thread timed out.").queue();
+                                    Config.clearCacheDirectory();
+                                    Wrapper.beingProcessed = false;
+                                } else {
+                                    try {
+                                        File newFilePNG = new File("cache/image" + newRandom + ".png");
+                                        ImageIO.write(processedImage, "png", newFilePNG);
+                                        message.delete().queue();
+                                        event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                                        Wrapper.beingProcessed = false;
+                                    } catch (Exception ignored) {
+                                        message.editMessage("Something went wrong with processing the image").queue();
+                                        Wrapper.beingProcessed = false;
+                                    }
+                                }
+                            });
                 } catch (Exception ex) {
                     message.editMessage("Something went wrong with processing the image").queue();
+                    Wrapper.beingProcessed = false;
                 }
 
             });
         }
+
         Config.clearCacheDirectory();
     }
 
+    private BufferedImage grayScaleImg(BufferedImage image) {
+        int width = image.getWidth();
+        int height = image.getHeight();
+        for (int y = 0; y < height; y++) {
+            for (int x = 0; x < width; x++) {
+                int p = image.getRGB(x, y);
+                int a = (p >> 24) & 0xff;
+                int r = (p >> 16) & 0xff;
+                int g = (p >> 8) & 0xff;
+                int b = p & 0xff;
+                int avg = (r + g + b) / 3;
+                p = (a << 24) | (avg << 16) | (avg << 8) | avg;
+                image.setRGB(x, y, p);
+            }
+        }
+        return image;
+    }
 
     @Override
     public String getHelp() {
