+++ src/main/java/me/comu/exeter/commands/image/PerspectiveImageCommand.java
@@ -1,5 +1,6 @@
 package me.comu.exeter.commands.image;
 
+
 import javafx.application.Platform;
 import javafx.embed.swing.JFXPanel;
 import javafx.embed.swing.SwingFXUtils;
@@ -22,18 +23,25 @@ import java.io.File;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Random;
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 
 public class PerspectiveImageCommand implements ICommand {
 
 
     @Override
     public void handle(List<String> args, GuildMessageReceivedEvent event) {
+        if (Wrapper.beingProcessed) {
+            event.getChannel().sendMessage("An image is already being processed, please wait.").queue();
+            return;
+        }
         if (event.getMessage().getAttachments().isEmpty()) {
             if (args.isEmpty()) {
                 event.getChannel().sendMessage("Please insert an image link to manipulate").queue();
                 return;
             }
+            Wrapper.beingProcessed = true;
             event.getChannel().sendMessage("`Processing Image...`").queue(message -> {
                 try {
                     int random = new Random().nextInt(1000);
@@ -41,16 +49,35 @@ public class PerspectiveImageCommand implements ICommand {
                     Wrapper.saveImage(args.get(0), "cache", "image" + random);
                     File file = new File("cache/image" + random + ".png");
                     BufferedImage image = ImageIO.read(file);
-                    File newFilePNG = new File("cache/image" + newRandom + ".png");
-                    ImageIO.write(distortImg(image), "png", newFilePNG);
-                    message.delete().queue();
-                    event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                      CompletableFuture.supplyAsync(() -> image)
+                            .thenApply(this::perspectiveImg)
+                            .completeOnTimeout(null, 10, TimeUnit.SECONDS)
+                            .thenAccept(processedImage -> {
+                                if (processedImage == null) {
+                                    message.editMessage("Processing thread timed out.").queue();
+                                    Config.clearCacheDirectory();
+                                    Wrapper.beingProcessed = false;
+                                } else {
+                                    try {
+                                        File newFilePNG = new File("cache/image" + newRandom + ".png");
+                                        ImageIO.write(processedImage, "png", newFilePNG);
+                                        message.delete().queue();
+                                        event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                                        Wrapper.beingProcessed = false;
+                                    } catch (Exception ignored) {
+                                        message.editMessage("Something went wrong with processing the image").queue();
+                                        Wrapper.beingProcessed = false;
+                                    }
+                                }
+                            });
                 } catch (Exception ignored) {
                     message.editMessage("Something went wrong with processing the image").queue();
+                    Wrapper.beingProcessed = false;
                 }
 
             });
         } else {
+            Wrapper.beingProcessed = true;
             event.getChannel().sendMessage("`Processing Image...`").queue(message -> {
                 try {
                     int random = new Random().nextInt(1000);
@@ -58,28 +85,30 @@ public class PerspectiveImageCommand implements ICommand {
                     Wrapper.saveImage(event.getMessage().getAttachments().get(0).getUrl(), "cache", "image" + random);
                     File file = new File("cache/image" + random + ".png");
                     BufferedImage image = ImageIO.read(file);
-                    double radius = 3.0 * image.getHeight();
-                    int xC = (int) radius, yC = (int) radius;
-                    BufferedImage imageB = new BufferedImage(xC * 2, yC, BufferedImage.TYPE_INT_ARGB);
-
-                    double r, i, j;
-                    for (int y = 0; y < imageB.getHeight(); y++) {
-                        for (int x = 0; x < imageB.getWidth(); x++) {
-                            r = Math.sqrt((xC - x) * (xC - x) + (yC - y) * (yC - y));
-                            i = (radius - r);
-                            j = (-3.0 * image.getWidth() / 2 * (xC - x)) / r + image.getWidth() / 2.0;
-                            if (i >= 0 && i < image.getHeight() && j >= 0 && j < image.getWidth()) {
-                                imageB.setRGB(x, y, image.getRGB((int) j, (int) i));
-                            }
-                        }
-                    }
-
-                    File newFilePNG = new File("cache/image" + newRandom + ".png");
-                    ImageIO.write(imageB, "png", newFilePNG);
-                    message.delete().queue();
-                    event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                    CompletableFuture.supplyAsync(() -> image)
+                            .thenApply(this::perspectiveImg)
+                            .completeOnTimeout(null, 10, TimeUnit.SECONDS)
+                            .thenAccept(processedImage -> {
+                                if (processedImage == null) {
+                                    message.editMessage("Processing thread timed out.").queue();
+                                    Config.clearCacheDirectory();
+                                    Wrapper.beingProcessed = false;
+                                } else {
+                                    try {
+                                        File newFilePNG = new File("cache/image" + newRandom + ".png");
+                                        ImageIO.write(processedImage, "png", newFilePNG);
+                                        message.delete().queue();
+                                        event.getChannel().sendFile(newFilePNG).queue(lol -> Config.clearCacheDirectory());
+                                        Wrapper.beingProcessed = false;
+                                    } catch (Exception ignored) {
+                                        message.editMessage("Something went wrong with processing the image").queue();
+                                        Wrapper.beingProcessed = false;
+                                    }
+                                }
+                            });
                 } catch (Exception ex) {
                     message.editMessage("Something went wrong with processing the image").queue();
+                    Wrapper.beingProcessed = false;
                 }
 
             });
@@ -87,7 +116,7 @@ public class PerspectiveImageCommand implements ICommand {
         Config.clearCacheDirectory();
     }
 
-    private static BufferedImage distortImg(BufferedImage image) {
+    private BufferedImage perspectiveImg(BufferedImage image) {
         new JFXPanel();
 
         final BufferedImage[] imageContainer = new BufferedImage[1];
@@ -145,7 +174,7 @@ public class PerspectiveImageCommand implements ICommand {
 
     @Override
     public String[] getAlias() {
-        return new String[]{"perspective", "perspectifyimage", "perspectifyimg", "perspectiveimage", "imageperspective", "imgperspective", "imgperspectify","imageperspectify"};
+        return new String[]{"perspective", "perspectifyimage", "perspectifyimg", "perspectiveimage", "imageperspective", "imgperspective", "imgperspectify", "imageperspectify"};
     }
 
     @Override
